Renames language visitors to Iso639-1 prefixes
