+++ src/main/java/com/progressiveaccess/cmlspeech/speech/en/EnAtomTable.java
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 /**
- * @file   EnglishAtomTable.java
+ * @file   EnAtomTable.java
  * @author Volker Sorge
  *         <a href="mailto:V.Sorge@progressiveaccess.com">Volker Sorge</a>
  * @date   Fri Jun 20 02:00:25 2014
@@ -34,11 +34,11 @@ import com.progressiveaccess.cmlspeech.speech.AbstractAtomTable;
  * Maps atom identifiers to their proper names.
  */
 
-public final class EnglishAtomTable extends AbstractAtomTable {
+public final class EnAtomTable extends AbstractAtomTable {
 
   private static final long serialVersionUID = 1L;
 
-  public EnglishAtomTable() {
+  public EnAtomTable() {
     this.put("Ac", "Actinium");
     this.put("Al", "Aluminum");
     this.put("Am", "Americium");
